+++ lib/watir-webdriver/generator/html.rb
@@ -1,9 +1,3 @@
-require "nokogiri"
-require "open-uri"
-require "pp"
-require "webidl"
-require "active_support/inflector"
-
 ActiveSupport::Inflector.inflections do |inflect|
   inflect.plural 'body', 'bodys'
   inflect.plural 'tbody', 'tbodys'
@@ -16,8 +10,6 @@ ActiveSupport::Inflector.inflections do |inflect|
   inflect.plural 'datalist', 'datalists'
 end
 
-require "watir-webdriver/html/util"
-require "watir-webdriver/html/visitor"
-require "watir-webdriver/html/idl_sorter"
-require "watir-webdriver/html/spec_extractor"
-require "watir-webdriver/html/generator"
+require "watir-webdriver/generator/html/generator"
+require "watir-webdriver/generator/html/spec_extractor"
+require "watir-webdriver/generator/html/visitor"